backend/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
lxml==4.6.3
pdf2image==1.15.1
Pillow==8.2.0
PyMuPDF==1.18.13
pytz==2021.1
SimpleIDML==1.1.2
sqlparse==0.4.1
suds-py3==1.4.4.1
